refactor: complete modular architecture with full functionality preservation
**Core Refactoring:**
- Split content.ts into 11 modular files maintaining 100% feature parity
- Reorganized popup.ts and background.ts into proper directory structure
- Updated webpack.config.js and tsconfig.json for new entry points

**New Monitors Created:**
- FormMonitor.ts: Form submission tracking with field analysis
- VisibilityMonitor.ts: Page visibility and interruption detection
- HoverMonitor.ts: Mouse hover pattern detection with debouncing

**Smart Assistant Integration:**
- Created smart-assistant-bridge.ts for page/extension communication
- Preserved all browser.runtime.onMessage handlers
- Maintained window.addEventListener message routing

**File Organization:**
src/
├── background/
│   ├── index.ts (main entry)
│   └── services/ (DataStorage, MLService, MessageRouter, StateManager)
├── content/
│   ├── index.ts (unified entry point)
│   ├── feature-extractor.ts (CSS selectors, event creation)
│   ├── smart-assistant-bridge.ts (extension communication)
│   └── monitors/ (7 specialized event monitors)
├── popup/index.ts (popup controller)
├── shared/ (types.ts, utils.ts, indexeddb-manager.ts)
├── smart-assistant/index.ts
└── workers/ml-worker.ts

**Verification:**
✅ All 28 original functions and classes preserved
✅ All 27 addEventListener calls maintained
✅ Smart Assistant initialization restored
✅ Browser message communication intact
✅ Build successful with zero errors
✅ Types and imports correctly updated
