+++ src/content/index.ts
@@ -1,9 +1,13 @@
 /// <reference path="../shared/types.ts" />
 import { createSynapseEvent } from './feature-extractor';
-import { sendToBackground, EventThrottler } from '../shared/utils';
+import { sendToBackground, EventThrottler, getModifierKeys } from '../shared/utils';
 import { setupScrollMonitoring } from './monitors/ScrollMonitor';
 import { setupFocusChangeMonitoring } from './monitors/FocusMonitor';
 import { setupClipboardMonitoring } from './monitors/ClipboardMonitor';
+import { setupFormSubmitMonitoring } from './monitors/FormMonitor';
+import { setupPageVisibilityMonitoring } from './monitors/VisibilityMonitor';
+import { setupMouseHoverMonitoring } from './monitors/HoverMonitor';
+import { setupSmartAssistantBridge } from './smart-assistant-bridge';
 import './monitors/TextInputAggregator';
 import './monitors/MouseTrajectoryMonitor';
 
@@ -11,7 +15,7 @@ declare var browser: any;
 
 const eventThrottler = new EventThrottler();
 
-// Basic click monitoring
+// Optimized click monitoring
 document.addEventListener('click', (event: MouseEvent) => {
   const element = event.target as HTMLElement;
   
@@ -21,22 +25,73 @@ document.addEventListener('click', (event: MouseEvent) => {
   });
 }, true);
 
-// Basic keyboard monitoring  
+// Optimized keyboard monitoring
 document.addEventListener('keydown', (event: KeyboardEvent) => {
   if (['Control', 'Shift', 'Alt', 'Meta'].includes(event.key)) {
     return;
   }
 
+  const modifierKeys = getModifierKeys(event);
+  
+  const isShortcut = modifierKeys.length > 0;
+  const isFunctionKey = event.key.startsWith('F') && event.key.length <= 3;
+  const isSpecialKey = ['Escape', 'Tab', 'Delete', 'Backspace', 'Insert', 'Home', 'End', 'PageUp', 'PageDown'].includes(event.key);
+  const isArrowKey = ['ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight'].includes(event.key);
+  
   const target = event.target as HTMLElement;
+  const isInputField = ['input', 'textarea'].includes(target.tagName.toLowerCase()) || target.contentEditable === 'true';
+  const isEnterInInput = event.key === 'Enter' && isInputField;
+  
+  if (!isShortcut && !isFunctionKey && !isSpecialKey && !isArrowKey && !isEnterInInput) {
+    return;
+  }
+
   eventThrottler.throttleEvent(event, () => {
-    const synapseEvent = createSynapseEvent('ui.keydown', target, event);
+    const synapseEvent = createSynapseEvent('ui.keydown', target, event, {
+      code: event.code,
+      modifier_keys: modifierKeys
+    });
+    
     sendToBackground(synapseEvent);
   });
 }, true);
 
-// Initialize all monitoring
-setupScrollMonitoring();
-setupFocusChangeMonitoring(); 
-setupClipboardMonitoring();
+function initializeAdvancedEventMonitoring(): void {
+  setupScrollMonitoring();
+  
+  window.addEventListener('beforeunload', () => {
+    // Cleanup throttlers if needed
+  });
+  
+  console.log('[Synapse] Advanced event monitoring initialized with throttling and debouncing');
+}
 
-console.log('[Synapse] Content script loaded with complete event monitoring suite.');
+function initializeAllEventMonitoring(): void {
+  initializeAdvancedEventMonitoring();
+  
+  setupFormSubmitMonitoring();
+  setupFocusChangeMonitoring();
+  setupPageVisibilityMonitoring();
+  setupMouseHoverMonitoring();
+  setupClipboardMonitoring();
+  
+  console.log('[Synapse] All event monitoring initialized including CLAUDE.md patterns');
+}
+
+// Initialize all event monitoring
+initializeAllEventMonitoring();
+
+// Add debug logging for missing events
+console.log('[Synapse] All event monitoring initialized. Expected events:');
+console.log('- user.scroll: scroll ≥20px');
+console.log('- ui.mouse_pattern: significance ≥0.02'); 
+console.log('- form.submit: form submit or button click');
+console.log('- ui.focus_change: focus in/out');
+console.log('- browser.page_visibility: visibility change');
+console.log('- ui.mouse_hover: hover ≥100ms');
+console.log('- ui.clipboard: copy/cut/paste');
+
+// Initialize smart assistant bridge
+setupSmartAssistantBridge();
+
+console.log('[Synapse] Content script loaded with complete event monitoring suite and smart assistant.');
