+++ src/shared/indexeddb-manager.ts
