+++ tsconfig.json
@@ -12,7 +12,8 @@
     "moduleResolution": "node",
     "baseUrl": "./src",
     "paths": {
-      "@core/*": ["core/*"],
+      "@core/*": ["background/services/*"],
+      "@shared/*": ["shared/*"],
       "@/*": ["./*"]
     }
   },
